remove provider and config.yml support
