+++ docs/source/usage/commandline_interface.rst
@@ -78,7 +78,6 @@ Build `style.json` from `style.yml`
     -u, --sprite-url [<sprite url>]                url to set as the sprite in style.json
     -i, --sprite-input [<icon input directory>]    directory path of icon source to build icons. The default <icon source> is `icons/`
     -o, --sprite-output [<icon output directory>]  directory path to output icon files. The default <icons destination> is the current directory
-    --provider [provider]                          your map service. e.g. `geolonia`
     -h, --help                                     display help for command
 
 Realtime editor on browser
@@ -92,7 +91,6 @@ Realtime editor on browser
     serve your map locally
 
     Options:
-    --provider [provider]                        your map service. e.g. `geolonia`
     -i, --sprite-input [<icon input directory>]  directory path of icon source to build icons. The default <icon
                                                  source> is `icons/`
     --sdf                                        Allows to use SDF sprite in charites
@@ -100,10 +98,6 @@ Realtime editor on browser
     -h, --help                                   display help for command
 
 Charites has three options for `serve` command.
-
-- ``--provider`` - `geolonia`, or `default`. When not specified, default or the value in the configuration file will be used.
-
-  - `geolonia` and `default` - the format linter runs against the MapLibre GL JS compatible specification.
 
 - ``--sprite-input`` - If you are building icon spritesheets with Charites, you can specify the directory of SVG files to compile here. See the ``build`` command for more information.
