+++ docs/source/usage/index.rst
@@ -6,6 +6,5 @@ Usage
   :maxdepth: 1
 
   commandline_interface
-  global_options
   examples
   example2
