--- provider/geolonia/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html id="charites-geolonia">
-  <head>
-    <link rel="stylesheet" href="/app.css" />
-    <title>Charites Live Preview</title>
-    <link href='https://unpkg.com/@watergis/maplibre-gl-legend@2.0.0/dist/maplibre-gl-legend.css' rel='stylesheet' />
-  </head>
-  <body>
-    <div class="overlay">
-      <label><input type="checkbox" id="showTileBoundaries">show Tile Boundaries</label>
-      <br>
-      <label><input type="checkbox" id="showCollisionBoxes">show Collision Boxes</label>
-      <br>
-      <label><input type="checkbox" id="showPadding">show Padding</label>
-    </div>
-    <div id="map"></div>
-    <script type="text/javascript" src="https://cdn.geolonia.com/v1/embed?geolonia-api-key=YOUR-API-KEY"></script>
-    <script src="https://unpkg.com/@watergis/maplibre-gl-legend@2.0.0/dist/maplibre-gl-legend.umd.js"></script>
-    <script type="text/javascript" src="/shared.js"></script>
-    <script type="text/javascript" src="/app.js"></script>
-  </body>
-</html>
