+++ test/convert.spec.ts
@@ -52,7 +52,7 @@ id: example
 
     const outJsonPath = path.join(tmp, 'converted-back.json')
     // This will throw an error if the outputted YAML was invalid
-    await build(yamlPath, outJsonPath, { provider: 'default' })
+    await build(yamlPath, outJsonPath, {})
     assert.isTrue(fs.existsSync(outJsonPath))
   })
